update code to centralize management of resort/region
There was duplicate logic for things like resort codes and names. This
made coding more difficult than needed. This adds the notion of a
region/resort type object that makes things a little easier to read.

It also lays the ground work for a new feature where we can report
*all* information to a user. This will allow them to detect when
new resorts/regions are added so that the android app can notify
them.