try to avoid race condition when creating cache file
There is no locking in place, so its possible for two perfectly
timed requests to both write to the cache file at the same time.

This change adds an flock in to prevent people from writing at the
same time. It also puts a shared lock so that the file won't be
overwritten while someone is reading it