add monitoring/testing support for server
We need some rudimentary monitoring support to ensure the server(s)
are responding and providing sane data.

I've updated test.sh to ensure a given list of required properties
are being reported by the server for each report module we have.
I also added some ability to help make the script more quiet or
verbose depending on what you need.

check-server.sh is a new script that is designed to be run from
cron. If run with the verbose flag it will always email a full
report of all data. If the verbose flag is not specified then it
will only send an email if the server doesn't respond or a given
property is missing.

The non-verbose check-server.sh run can be thought of as our
monitoring. If all goes well, nothing happens. However, if something
goes wrong you'll get an email that looks like:

WakeMeSki Server Error!!!
= testing: http://bettykrocks.com/skireport/nwac_report.php?location=OSOALP
ERROR: missing required property(cache.found)