Add ability to generate usage statistics
This adds some rudimentary support for tracking usage statistics.
Basically we append "hit" information to a log file. Each PHP file
can call a utility function to handle the logic. I wrote a simple
parser to generate some very basic statistics for now. This can
be enhanced in the future if needed.