add support for spain, improve france
A user requested support for Cerler Spain. I added that and
went ahead did the rest of the resorts we can support for Spain.

Updated France with a few more resort.

Attempted to configure websites and GPS coordinates for the
France/Spain reports. These are only as accurate as google and
wikipedia could help me figure out

Update eu_report.php to handle reports that provide a resort
url and/or lat/long