fix vermont reporting error
snow.daily isn't reported today because of a format change
in the report's HTML. I've updated the logic to use the
find_int utility function.

Since there is no fresh snow in Vermont, I'll have to check
back tomorrow to make sure it still works when there is
fresh snow.