fix vermont ski reports
The format of the fresh snow reporting changed in the
reports HTML. They also changed the wording of the mountain
names.

I fixed the logic to find fresh snow. I also made the logic
for finding the HTML for a given resort more flexible so its
not going to break if they rename "Jay Peak" to "Jay Peak Resort"