update vermont to have more flexible parsing
The resorts in Vermont sometimes report snow measurements as a
single value like: 3". Sometimes the do ranges like: 1-3".
This adds the flexibility to handle either.

I've also added a new property "snow.making". It just says if
the resort is making snow. The client doesn't yet have support,
but I thought over time we might be able to support this property.

The last thing I've created is a new "examples" directory. This
has copies of reports that we parse for a location. This should
allow us to compare reports and get an idea of how things have
changed. It should help maintainence and debugability