fix bugs in BC reports
The weather wasn't showing up because we were reporting weather
with props:
 forecast.*
It needs to be:
 weather.forecast.*
We also were using 1-based indices rather than 0-based

Latitude/Longitude weren't showing up because common.inc was done
so that it only adds them to the report when you handle US based
weather reports.

bc.inc had the longitude wrong for Whistler

Also - the "cloudy" entry for weather icon is wrong. The android
client needs this to be "ovc"