Change polling rate before sunrise
We've had one user complaint about slow server refresh in early
morning hours.  I've also noticed this when checking server status
early (before 6) in the morning.  In at least one case the feed
had been updated but the application still showed report data from
the day before.

This commit attempts to modify the polling interval when all of
the following conditions are true:
 * The report is yesterday's report
 * The time is morning (before sunset) at the resort latitude and
	longitude.

When both conditions are true, the polling interval changes to
10 minutes for cache refresh instead of the default 2 hours
as set by cache_available.

Added methods in common.inc to calculate and compare cache date
values to the current time and sunrise time.  Modify each
invocation of cache_available to pass the resort object to allow
latitude and longitude values to be used in calculation of sunrise
time.

Interestingly when I look at the debug logs it seems gmdate does
not return the time in GMT (I see a MST value) however I don't think
it will matter what the offset is since both gmdate and date_sunrise
do appear to use the same offset and sunrise values appear to be
reasonable wrt MST.

I've noticed the reports from nwac appear to have a broken report
date when I made this change.